requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
et-xmlfile==1.1.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.0
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.2
pypdf==3.16.0
python-dateutil==2.8.2
pytz==2023.3.post1
rawpy==0.24.0
six==1.16.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2023.3
uvicorn==0.34.0
